% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursoriality_theropod.R
\name{cursoriality_theropod}
\alias{cursoriality_theropod}
\title{Calculate Cursoriality Parameters for Theropods}
\usage{
cursoriality_theropod(weight_kg)
}
\arguments{
\item{weight_kg}{Numeric. The body mass of the theropod in kilograms.}
}
\value{
A list with two elements:
\describe{
\item{Optimal_cadence}{Numeric. The optimal cadence for the given body mass (steps per second).}
\item{Optimal_stride_ratio}{Numeric. The optimal stride ratio for the given body mass.}
}
}
\description{
This function calculates the optimal cadence and stride ratio for theropods based on their body mass.
}
\details{
The function categorizes theropods into weight ranges and assigns optimal cadence and stride ratio values based on these ranges.
These values are derived from biomechanical principles and empirical studies of extant cursorial animals.

The cadence represents the number of steps taken per second, while the stride ratio is a dimensionless value
that represents the relationship between stride length and body proportions. For theropods heavier than 10,000 kg,
the function assumes minimal cursoriality with cadence and stride ratio values of 1.
It is crucial to provide the body mass in kilograms to ensure accurate parameter calculations.
}
\note{
This function is designed for paleontological research and educational purposes. The values it provides are estimates
based on general models and may not fully represent specific species.
}
\examples{
cursoriality_theropod(500)

}
\references{
Coombs, W. P. (1978). "Theoretical Aspects of Cursorial Adaptations in Dinosaurs."
Alexander, R. M. (1976). "Estimates of Speeds of Dinosaurs." Nature, 261:129–130.
Molina and Larramendi (2016). "Dinosaur Facts and Figures - The Theropods and Other Dinosauriformes"
}
\author{
Juan Ignacio Castillo
}
